+++ src/main/java/rentalstore/NewReleaseMovie.java
@@ -8,7 +8,7 @@ public class NewReleaseMovie extends Movie {
         return day * 3.0;
     }
 
-    protected int getFrequentRenterPoints(int day) {
+    protected double getFrequentRenterPoints(int day) {
         if(day>1){
             return 2;
         }
