+++ src/main/java/rentalstore/Rental.java
@@ -21,7 +21,7 @@ public class Rental {
         return movie.getAmount(dayRented);
     }
 
-    public int getFrequentRenterPoints() {
+    public double getFrequentRenterPoints() {
         return movie.getFrequentRenterPoints(dayRented);
     }
 }
